separate migration code, add cli for missing alt_ids
